test(patina): move permitted_contents tests to a sibling file
